Add support for signing in with a Google account.
This makes it possible to select a reader identity provided by the
reader backend which depends on what Google account was signed in
with.

This requires an Google Cloud Console project to be set up as detailed
in https://developer.android.com/identity/sign-in/credential-manager-siwg
and this includes references (server clientId) to such a service set up
for testing. The APKs deployed on apps.multipaz.org will use a separate
service with another clientId so make it possibel to easily configure
this.

For now this is only implemented for Android but should be
straightforward to implement on iOS as well. This will be done in a
future commit.

Test: Manually tested.
Signed-off-by: David Zeuthen <zeuthen@google.com>